// src/app/Models/ManInfoClass.ts
// Generated by https://quicktype.io

export interface ManInfoClass {
    EmpName:   string;
    DeptID:    string;
    DeptName:  string;
    DeptPath:  string;
    PosID:     string;
    PosName:   string;
    Auth:      boolean;
    Email:     string;
    MainMan:   boolean;
    ChiefCode: string;
    RoleEmp:   string;
    RoleID:    string;
    RoleName:  string;
    EmpID:     string;
}