// src/app/Models/GetCheckAgentClass.ts
// Generated by https://quicktype.io

export class GetCheckAgentClass {
    AutoKey:            number;
    Key:                string;
    CheckAgentDept:     any[];
    CheckAgentFlowTree: any[];
    EmpAgentDate:       EmpAgentDate[];
    RoleIdSource:       string;
    RoleNameSource:     string;
    EmpIdSource:        string;
    EmpNameSource:      string;
    RoleIdTarget:       any;
    RoleNameTarget:     any;
    EmpIdTarget:        string;
    EmpNameTarget:      string;
    DateB:              string;
    DateE:              string;
    IsValid:            boolean;
    Sort:               number;
    KeyMan:             string;
    KeyDate:            string;
    EmpName:            any;
    DeptID:             any;
    DeptName:           any;
    DeptPath:           any;
    PosID:              any;
    PosName:            any;
    Auth:               boolean;
    Email:              any;
    MainMan:            boolean;
    ChiefCode:          any;
    RoleEmp:            any;
    RoleID:             any;
    RoleName:           any;
    EmpID:              any;
}

export interface EmpAgentDate {
    AutoKey: number;
    EmpID:   string;
    DateB:   string;
    DateE:   string;
    KeyMan:  string;
    KeyDate: string;
    IsValid: boolean;
}